[clients] introduce feature-flag "authorized"
all the APIs whose endpoints might need authorization are hidden behind
a feature-flag named "authorized".
Disabling the flag makes the clients wasm-compatible
