+++ crates/bcr-wdc-quote-service/Cargo.toml
@@ -14,9 +14,9 @@ async-trait.workspace = true
 axum = {workspace = true, features = ["macros"]}
 axum-test = {workspace = true, optional = true}
 bcr-ebill-core.workspace = true
-bcr-wdc-key-client = {workspace = true}
+bcr-wdc-key-client = {workspace = true, features = ["authorized"]}
 bcr-wdc-utils = {workspace = true}
-bcr-wdc-treasury-client = {workspace = true}
+bcr-wdc-treasury-client = {workspace = true, features = ["authorized"]}
 bcr-wdc-webapi = {workspace = true}
 bitcoin.workspace = true
 borsh.workspace = true
