+++ crates/bcr-wdc-treasury-service/Cargo.toml
@@ -8,8 +8,8 @@ license.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
 axum = {workspace = true, features = ["macros"]}
-bcr-wdc-key-client = {workspace = true}
-bcr-wdc-quote-client = {path = "../bcr-wdc-quote-client"}
+bcr-wdc-key-client = {workspace = true, features = ["authorized"]}
+bcr-wdc-quote-client = {workspace = true, features = ["authorized"]}
 bcr-wdc-swap-client = {workspace = true}
 bcr-wdc-utils = {workspace = true}
 bcr-wdc-webapi = { workspace = true}
