+++ crates/bcr-wdc-treasury-service/src/lib.rs
@@ -11,6 +11,7 @@ mod credit;
 mod debit;
 mod error;
 mod persistence;
+mod web;
 // ----- local imports
 
 type ProdCreditRepository = persistence::surreal::CreditRepository;
@@ -82,18 +83,19 @@ impl AppController {
 }
 
 pub fn routes(app: AppController) -> Router {
+    let web = Router::new().route("/v1/redeem", post(web::redeem));
     Router::new().nest(
         "/v1/admin/treasury",
         Router::new()
             .route("/credit/generate_blinds", post(admin::generate_blinds))
             .route("/credit/store_signatures", post(admin::store_signatures))
             .route("/credit/balance", get(admin::crsat_balance))
-            .route("/debit/redeem", post(admin::redeem))
             .route("/debit/balance", get(admin::sat_balance))
             .route(
                 "/debit/request_to_mint_from_ebill",
                 post(admin::request_mint_from_ebill),
             )
+            .merge(web)
             .with_state(app),
     )
 }
