+++ crates/bcr-wdc-utils/Cargo.toml
@@ -7,6 +7,7 @@ license.workspace = true
 
 [features]
 test-utils = ["rand"]
+auth = ["reqwest", "serde", "tokio", "tracing"]
 
 
 [dependencies]
@@ -19,12 +20,12 @@ cdk-common = {workspace = true}
 chrono.workspace = true
 itertools = {workspace = true}
 rand = {workspace = true, optional  = true}
-reqwest = {workspace = true, features = ["json"]}
+reqwest = {workspace = true, features = ["json"], optional = true}
 secp256k1 = {workspace = true, features = ["global-context"]}
-serde = {workspace = true, features = ["derive"]}
+serde = {workspace = true, features = ["derive"], optional = true}
 thiserror.workspace = true
-tokio = {workspace = true, features = ["time", "rt"]}
-tracing = {workspace = true}
+tokio = {workspace = true, features = ["time", "rt"], optional = true}
+tracing = {workspace = true, optional = true}
 url.workspace = true
 uuid.workspace = true
