+++ crates/bcr-wdc-utils/src/lib.rs
@@ -2,6 +2,7 @@
 // ----- extra library imports
 // ----- local modules
 pub mod borsh;
+#[cfg(feature = "auth")]
 pub mod client;
 pub mod keys;
 pub mod signatures;
